Move getAddress func to chain-util
