+++ tx-pool/transaction.js
@@ -37,7 +37,7 @@ class Transaction {
       address = txBody.address;
       skipVerif = true;
     } else {
-      address = Transaction.getAddress(hash.slice(2), signature);
+      address = ChainUtil.getAddress(hash.slice(2), signature);
     }
     const createdAt = Date.now();
     return new Transaction(txBody, signature, hash, address, skipVerif, createdAt);
@@ -82,25 +82,6 @@ class Transaction {
 
   toString() {
     return JSON.stringify(this, null, 2);
-  }
-
-  /**
-   * Gets address from hash and signature.
-   */
-  static getAddress(hash, signature) {
-    let address = '';
-    try {
-      const sigBuffer = ainUtil.toBuffer(signature);
-      const len = sigBuffer.length;
-      const lenHash = len - 65;
-      const {r, s, v} = ainUtil.ecSplitSig(sigBuffer.slice(lenHash, len));
-      const publicKey = ainUtil.ecRecoverPub(Buffer.from(hash, 'hex'), r, s, v);
-      address = ainUtil.toChecksumAddress(ainUtil.bufferToHex(
-          ainUtil.pubToAddress(publicKey, publicKey.length === 65)));
-    } catch (err) {
-      logger.error(err);
-    }
-    return address;
   }
 
   /**
